Moved BooleanButton and COmandScheduler Wrapper
