+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/BooleanButton.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.subsystems;
+package org.firstinspires.ftc.teamcode.util;
 
 import androidx.annotation.NonNull;
